# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateAirdropRequest:
    properties:
      category:
        example: rumored
        type: string
      chain:
        example: Ethereum
        type: string
      name:
        example: zkSync
        type: string
      notes:
        example: Bridge weekly
        type: string
      priority:
        example: high
        type: string
      url:
        example: https://zksync.io
        type: string
    type: object
  handler.CreateTaskRequest:
    properties:
      description:
        example: Swap ETH to USDC
        type: string
      frequency:
        example: weekly
        type: string
    type: object
...
  handler.UpdateAccountAirdropRequest:
    properties:
      notes:
        example: Focus on bridging
        type: string
      status:
        example: active
        type: string
    type: object
...
  model.Account:
    properties:
      account_airdrops:
        items:
          $ref: '#/definitions/model.AccountAirdrop'
        type: array
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      wallets:
        description: Relations
        items:
          $ref: '#/definitions/model.Wallet'
        type: array
    type: object
  model.AccountAirdrop:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/model.Account'
        description: Relations
      account_id:
        type: integer
      airdrop:
        $ref: '#/definitions/model.Airdrop'
      airdrop_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      status:
        description: active, completed, paused
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      updated_at:
        type: string
    type: object
  model.Airdrop:
    properties:
      category:
        type: string
      chain:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      priority:
        type: string
      status:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: integer
    type: object
  model.Task:
    properties:
      account_airdrop:
        allOf:
        - $ref: '#/definitions/model.AccountAirdrop'
        description: Relations
      account_airdrop_id:
        type: integer
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      frequency:
        description: once, daily, weekly, monthly
        type: string
      gas_spent:
        type: number
      id:
        type: integer
      is_completed:
        type: boolean
      next_due:
        type: string
      tx_hash:
        type: string
    type: object
...
  model.Wallet:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/model.Account'
        description: Relations
      account_id:
        type: integer
      address:
        type: string
      chain:
        type: string
      created_at:
        type: string
      id:
        type: integer
      label:
        type: string
      user_id:
        type: integer
    type: object
host: airdrop-tracker-api-v1.fly.dev
info:
  contact:
    email: support@airdrop-tracker.local
    name: Bita
  description: Personal airdrop task management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Airdrop Tracker API
  version: "1.0"
paths:
  /api/account-airdrops/{id}:
    get:
      description: Get a specific account-airdrop with tasks
      parameters:
      - description: Account Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AccountAirdrop'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account-airdrop detail
      tags:
      - AccountAirdrops
    put:
      consumes:
      - application/json
      description: Update status or notes for an account-airdrop assignment
      parameters:
      - description: Account Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAccountAirdropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AccountAirdrop'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account-airdrop
      tags:
      - AccountAirdrops
  /api/account-airdrops/{id}/tasks:
    get:
      description: Get all tasks for a specific account-airdrop
      parameters:
      - description: Account Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tasks for account-airdrop
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Add new task to an account-airdrop
      parameters:
      - description: Account Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - Tasks
...
  /api/airdrops:
    get:
      description: Get all global airdrops for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Airdrop'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all airdrops
      tags:
      - Airdrops
    post:
      consumes:
      - application/json
      description: Add new global airdrop to the catalog
      parameters:
      - description: Airdrop data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAirdropRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Airdrop'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create airdrop
      tags:
      - Airdrops
  /api/airdrops/{id}:
    delete:
      description: Remove airdrop from the global catalog
      parameters:
      - description: Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete airdrop
      tags:
      - Airdrops
    get:
      description: Get single airdrop by ID
      parameters:
      - description: Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Airdrop'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get airdrop detail
      tags:
      - Airdrops
    put:
      consumes:
      - application/json
      description: Update existing airdrop
      parameters:
      - description: Airdrop ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAirdropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Airdrop'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update airdrop
      tags:
      - Airdrops
...
  /api/tasks/{id}:
    delete:
      description: Remove a task permanently
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - Tasks
  /api/tasks/{id}/complete:
    put:
      description: Mark a task as done
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark task as completed
      tags:
      - Tasks
  /api/tasks/{id}/reset:
    put:
      description: Reset a completed task back to pending
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reset task
      tags:
      - Tasks
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"